+++ SpaceBattle.Lib/Uobject/Adapter.cs
@@ -1,24 +1,33 @@
-// namespace SpaceBattle.Lib;
+namespace SpaceBattle.Lib;
+//System refllection;
 //Activator
-//Imovable movable =IoC.Resolve<IMovable>("GenerateAdapter", typeOf(IMovable), object);  стратегия 
-//System refllection
-//Написать стратегию для генерации адаптеров
-// Тред локал айм трап айм трап вылизал её с головы до ног оу оу я снова обогнал, снова обогнал предлагали 10, забрал 10 по 10
-// public class MovableAdapter: IMovable
-// {
-//     Vector Position
-//     {
-//         get{
-//             return (Vector)IoC.Resolve<object>("UObject.getProperty", this.obj, "Position");
-//         }
-//         set (object value){
-//             IoC.Resolve<ICommand>("UObject.setProperty", this.obj, "Position");
-//         }
-//     }
-//     Vector Velocity
-//     {
-//         get{
-//             return (Vector)IoC.Resolve<object>("UObject.getProperty", this.obj, "Velocity");
-//         }
-//     }
-// }
+//Imovable movable =IoC.Resolve<IMovable>("GenerateAdapter", typeOf(IMovable), object);  стратегия
+
+public class MovableAdapter : IMovable
+{
+    private IUObject obj;
+    MovableAdapter(IUObject obj)
+    {
+        this.obj = obj;
+    }
+    public Vector Position
+    {
+        get
+        {
+            return (Vector)IoC.Resolve<object>("UObject.getProperty", this.obj, "Position");
+        }
+        set
+        {
+            IoC.Resolve<ICommand>("UObject.setProperty", this.obj, "Position", value);
+        }
+    }
+    public Vector Velocity
+    {
+        get
+        {
+            return (Vector)IoC.Resolve<object>("UObject.getProperty", this.obj, "Velocity");
+        }
+    }
+
+
+}
